refactor: allow for ssr for better SEO
