+++ src/app/(dashboard)/layout.tsx
@@ -1,0 +1,23 @@
+"use server"
+
+import Header from "@/components/layout/header/context";
+import Navigation from "@/components/layout/nav/context";
+ 
+export default function DashboardLayout({
+    children,
+  }: Readonly<{
+    children: React.ReactNode;
+  }>) {
+  
+    return (
+        <>
+            {/* navigation / bav bar code here which uses dashboard.navigation for the value */}
+            <Navigation/>
+            <Header/>
+            <main>
+                {children}
+            </main>
+        </>
+    );
+}
+
