+++ src/app/(dashboard)/loading.tsx
@@ -1,0 +1,3 @@
+export default function DashboardLoading () {
+    return <div>Loading...</div>
+}
