+++ src/app/(dashboard)/quiz/[id]/page.tsx
