+++ src/app/(dashboard)/student/[id]/page.tsx
