+++ src/components/layout/header/context.tsx
@@ -1,0 +1,7 @@
+"use client"
+
+import HeaderBar from "./bar";
+
+export default function Header() {
+    return <HeaderBar name={""} isLoggedIn={false} />
+}
